# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vare AI Backend API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeneratedAt {
  generated_at: string;
}

export type ErrorResponseDetails = { [key: string]: unknown };

export interface ErrorResponse {
  error: string;
  code: string;
  details?: ErrorResponseDetails;
}

export type MerchantOnboardingStatus =
  (typeof MerchantOnboardingStatus)[keyof typeof MerchantOnboardingStatus];

export const MerchantOnboardingStatus = {
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Merchant {
  id: string;
  companyName: string;
  contactFirstName?: string | null;
  contactLastName?: string | null;
  contactEmail: string;
  contactPhone?: string | null;
  estimatedSkuCount?: string | null;
  primaryVertical?: string | null;
  magentoVersion?: string | null;
  hostingEnvironment?: string | null;
  erpSystem?: string | null;
  pimSystem?: string | null;
  /**
   * @minimum 0
   * @maximum 4
   */
  complexityScore: number;
  /**
   * @minimum 1
   * @maximum 10
   */
  onboardingPhase: number;
  onboardingStatus: MerchantOnboardingStatus;
  apiKey?: string | null;
  sandboxMode: boolean;
  isLive: boolean;
  createdAt?: string | null;
  updatedAt?: string | null;
}

export interface CreateMerchantRequest {
  /** @minLength 1 */
  companyName: string;
  contactFirstName?: string;
  contactLastName?: string;
  contactEmail: string;
  contactPhone?: string;
  estimatedSkuCount?: string;
  primaryVertical?: string;
  magentoVersion?: string;
  hostingEnvironment?: string;
  erpSystem?: string;
  pimSystem?: string;
  sandboxMode?: boolean;
}

export interface UpdateMerchantRequest {
  companyName?: string;
  contactFirstName?: string;
  contactLastName?: string;
  contactEmail?: string;
  contactPhone?: string;
  estimatedSkuCount?: string;
  primaryVertical?: string;
  magentoVersion?: string;
  hostingEnvironment?: string;
  erpSystem?: string;
  pimSystem?: string;
  sandboxMode?: boolean;
}

export type ComplexityResponseLabel =
  (typeof ComplexityResponseLabel)[keyof typeof ComplexityResponseLabel];

export const ComplexityResponseLabel = {
  Simple: "Simple",
  Standard: "Standard",
  Moderate: "Moderate",
  Complex: "Complex",
  Enterprise: "Enterprise",
} as const;

export type ComplexityResponseFactors = {
  hasEnterpriseEdition?: boolean | null;
  hasErp?: boolean;
  hasPim?: boolean;
  largeSkuCount?: boolean;
};

export interface ComplexityResponse {
  merchantId: string;
  /**
   * @minimum 0
   * @maximum 4
   */
  complexityScore: number;
  label: ComplexityResponseLabel;
  factors: ComplexityResponseFactors;
}

export type GenerateKeyResponseMode =
  (typeof GenerateKeyResponseMode)[keyof typeof GenerateKeyResponseMode];

export const GenerateKeyResponseMode = {
  live: "live",
  test: "test",
} as const;

export interface GenerateKeyResponse {
  apiKey?: string | null;
  mode: GenerateKeyResponseMode;
  message: string;
}

export type SystemHealthResponseChecks = { [key: string]: unknown };

export interface SystemHealthResponse {
  status: string;
  checks: SystemHealthResponseChecks;
  timestamp: string;
}

export type HealthCheck200 = {
  data: HealthStatus;
  generated_at: string;
};

export type CreateMerchant201 = {
  data?: Merchant;
  generated_at?: string;
};

export type GetMerchant200 = {
  data?: Merchant;
  generated_at?: string;
};

export type UpdateMerchant200 = {
  data?: Merchant;
  generated_at?: string;
};

export type GetMerchantComplexity200 = {
  data?: ComplexityResponse;
  generated_at?: string;
};

export type GenerateApiKeyBodyMode =
  (typeof GenerateApiKeyBodyMode)[keyof typeof GenerateApiKeyBodyMode];

export const GenerateApiKeyBodyMode = {
  live: "live",
  test: "test",
} as const;

export type GenerateApiKeyBody = {
  mode?: GenerateApiKeyBodyMode;
};

export type GenerateApiKey200 = {
  data?: GenerateKeyResponse;
  generated_at?: string;
};

export type SaveConnectionBody = {
  storeUrl: string;
  consumerKey?: string;
  consumerSecret?: string;
  accessToken?: string;
  accessTokenSecret?: string;
  apiUser?: string;
  apiKeyM1?: string;
};

export type SaveConnection201Data = { [key: string]: unknown };

export type SaveConnection201 = {
  data?: SaveConnection201Data;
  generated_at?: string;
};

export type TestConnection200DataStoreViewsItem = { [key: string]: unknown };

export type TestConnection200Data = {
  success: boolean;
  storeName?: string | null;
  version?: string | null;
  currency?: string | null;
  locale?: string | null;
  storeViews?: TestConnection200DataStoreViewsItem[] | null;
  latencyMs?: number | null;
  error?: string | null;
  errorCode?: string | null;
};

export type TestConnection200 = {
  data?: TestConnection200Data;
  generated_at?: string;
};

export type ConnectionHealth200Data = {
  success: boolean;
  apiLatencyMs?: number | null;
  catalogEndpoint?: boolean | null;
  inventoryEndpoint?: boolean | null;
  apiHealthPct?: number | null;
  error?: string | null;
};

export type ConnectionHealth200 = {
  data?: ConnectionHealth200Data;
  generated_at?: string;
};

export type ListStoreViews200DataItem = { [key: string]: unknown };

export type ListStoreViews200 = {
  data?: ListStoreViews200DataItem[];
  generated_at?: string;
};

export type UpdateStoreViewSelectionsBody = {
  selected: string[];
};

export type UpdateStoreViewSelections200DataItem = { [key: string]: unknown };

export type UpdateStoreViewSelections200 = {
  data?: UpdateStoreViewSelections200DataItem[];
  generated_at?: string;
};

export type ConfigureSyncFiltersBody = {
  productTypes?: string[];
  status?: string[];
  visibility?: string[];
  categoryIds?: number[];
  attributes?: string[];
};

export type ConfigureSyncFilters200Data = { [key: string]: unknown };

export type ConfigureSyncFilters200 = {
  data?: ConfigureSyncFilters200Data;
  generated_at?: string;
};

export type StartSyncBodyType =
  (typeof StartSyncBodyType)[keyof typeof StartSyncBodyType];

export const StartSyncBodyType = {
  full: "full",
  delta: "delta",
} as const;

export type StartSyncBody = {
  type?: StartSyncBodyType;
};

export type StartSync202Data = {
  id: string;
  status: string;
  jobType: string;
};

export type StartSync202 = {
  data?: StartSync202Data;
  generated_at?: string;
};

export type GetSyncStatusParams = {
  jobId?: string;
};

export type GetSyncStatus200Data = {
  jobId: string;
  status: string;
  totalRecords?: number | null;
  processedRecords?: number | null;
  errorCount?: number | null;
};

export type GetSyncStatus200 = {
  data?: GetSyncStatus200Data;
  generated_at?: string;
};

export type PauseSyncBody = {
  jobId: string;
};

export type PauseSync200 = { [key: string]: unknown };

export type CancelSyncBody = {
  jobId: string;
};

export type CancelSync200 = { [key: string]: unknown };

export type GetSyncSummary200 = { [key: string]: unknown };

export type GetSyncErrorsParams = {
  jobId?: string;
  page?: number;
  /**
   * @maximum 100
   */
  limit?: number;
};

export type GetSyncErrors200 = { [key: string]: unknown };

export type SystemHealthCheck200 = {
  data?: SystemHealthResponse;
  generated_at?: string;
};