# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vare AI Backend API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Merchant API key (vare_live_sk_... or vare_test_sk_...)
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
      required:
        - error
        - code
...
paths:
...
  /onboarding/connect/test:
    post:
      operationId: testConnection
      tags: [connect]
      summary: Test Magento API connection
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                      storeName:
                        type: string
                        nullable: true
                      version:
                        type: string
                        nullable: true
                      currency:
                        type: string
                        nullable: true
                      locale:
                        type: string
                        nullable: true
                      storeViews:
                        type: array
                        items:
                          type: object
                        nullable: true
                      latencyMs:
                        type: integer
                        nullable: true
                      error:
                        type: string
                        nullable: true
                      errorCode:
                        type: string
                        nullable: true
                    required:
                      - success
                  generated_at:
                    type: string
                    format: date-time
        "400":
          description: No connection configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...